[Removed] forward declarations, moved them to fwd file
