+++ engine/reflection/public/member/field.hpp
@@ -2,35 +2,28 @@
 #include <common.hpp>
 
 #include <cstddef>
+#include <reflect_fwd.hpp>
 
-namespace reflect
+namespace reflect::detail
 {
-class Value;
-class Type;
-
-namespace detail
+class Field
 {
-    class ReflectFactory;
+public:
+    template <typename Class, typename Member> using MemberPointer = Member Class::*;
+    template <typename Class, typename Member>
+    Field(ReflectFactory& type_store, MemberPointer<Class, Member> mem_ptr);
+    ~Field();
 
-    class Field
-    {
-    public:
-        template <typename Class, typename Member> using MemberPointer = Member Class::*;
-        template <typename Class, typename Member>
-        Field(ReflectFactory& type_store, MemberPointer<Class, Member> mem_ptr);
-        ~Field();
+    NON_MOVABLE(Field);
+    NON_COPYABLE(Field);
 
-        NON_MOVABLE(Field);
-        NON_COPYABLE(Field);
+    NO_DISCARD void* rawAccess(void* ptr) const;
+    NO_DISCARD const Type* getType() const;
 
-        NO_DISCARD void* rawAccess(void* ptr) const;
-        NO_DISCARD const Type* getType() const;
-
-    private:
-        size_t offset {};
-        const Type* type {};
-    };
-}
+private:
+    size_t offset {};
+    const Type* type {};
+};
 }
 
 #include <member/field_impl.hpp>
