+++ engine/reflection/public/reflect_fwd.hpp
@@ -1,0 +1,16 @@
+namespace reflect
+{
+class Value;
+class ValueRef;
+class Type;
+class ArgumentList;
+class ParameterList;
+namespace detail
+{
+    class Constructor;
+    class ReflectFactory;
+    class Destructor;
+    class Field;
+    class Method;
+}
+}
